+++ src/forge/stop.py
@@ -5,8 +5,9 @@ import sys
 import boto3
 
 from . import REQUIRED_ARGS
-from .parser import add_basic_args, add_general_args, add_env_args
-from .common import ec2_ip, get_ip, set_boto_session, get_nlist
+from .parser import add_basic_args, add_general_args, add_env_args, add_job_args, add_action_args
+from .common import ec2_ip, get_ip, get_nlist
+from .configuration import Configuration
 
 logger = logging.getLogger(__name__)
 
@@ -22,6 +23,8 @@ def cli_stop(subparsers):
     parser = subparsers.add_parser('stop', description='Stop an on-demand EC2')
     add_basic_args(parser)
     add_general_args(parser)
+    add_job_args(parser, suppress=True)
+    add_action_args(parser, suppress=True)
     add_env_args(parser)
 
     REQUIRED_ARGS['stop'] = ['name',
@@ -29,19 +32,16 @@ def cli_stop(subparsers):
                              'forge_env']
 
 
-def stop_fleet(n_list, config):
+def stop_fleet(n_list, config: Configuration):
     """stops each fleet in n_list
 
     Parameters
     ----------
     n_list : list
         List of fleet names
-    config : dict
+    config : Configuration
         Forge configuration data
     """
-    profile = config.get('aws_profile')
-    region = config.get('region')
-    set_boto_session(region, profile)
     client = boto3.client('ec2')
 
     details = {n: ec2_ip(n, config) for n in n_list}
@@ -62,20 +62,21 @@ def stop_fleet(n_list, config):
             client.stop_instances(InstanceIds=[uid])
 
 
-def stop(config):
+def stop(config: Configuration):
     """stops a running on-demand EC2 instance
 
     Parameters
     ----------
-    config : dict
+    config : Configuration
         Forge configuration data
     """
-    market = config.get('market')
+    market = config.market
 
     if 'spot' in market:
         logger.error('Master or worker is a spot instance; you cannot stop a spot instance')
         # sys.exit(1)  # ToDo: Should we change the tests to reflect an exit or allow it to continue?
 
-    n_list = get_nlist({**config, 'rr_all': True})
+    config.rr_all = True
 
+    n_list = get_nlist(config)
     stop_fleet(n_list, config)