+++ tests/test_create.py
@@ -8,13 +8,24 @@ import pytest
 from botocore.exceptions import ClientError
 
 from forge import create
+from forge.configuration import Configuration
+
+
+BASE_CONFIG = {
+    'region': 'us-east-1',
+    'ec2_amis': {},
+    'ec2_key': '',
+    'forge_env': 'dev',
+    'forge_pem_secret': '',
+    'job': 'create'
+}
 
 
 @mock.patch('forge.create.get_instance_details')
 @mock.patch('forge.create.search_and_create')
 def test_create_single(mock_search_create, mock_get_instance_details):
     """Test entry-point for creation of single instance."""
-    config = {'service': 'single', 'aws_az': 'us-east-1a'}
+    config = Configuration(**{**BASE_CONFIG, 'service': 'single', 'aws_az': 'us-east-1a'})
     mock_get_instance_details.return_value = {'single': {}}
     create.create(config)
     mock_search_create.assert_called_once_with(config, 'single', {})
@@ -24,7 +35,7 @@ def test_create_single(mock_search_create, mock_get_instance_details):
 @mock.patch('forge.create.search_and_create')
 def test_create_cluster_master_workers(mock_search_create, mock_get_instance_details):
     """Test entry-point for creation of cluster with master and workers."""
-    config = {'service': 'cluster', 'aws_az': 'us-east-1a'}
+    config = Configuration(**{**BASE_CONFIG, 'service': 'cluster', 'aws_az': 'us-east-1a'})
     mock_get_instance_details.return_value = {'cluster-master': {}, 'cluster-worker': {}}
     create.create(config)
     mock_search_create.assert_has_calls([