+++ CHANGELOG.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3]
+
+### Changed
+- **Create** - AMI ID restrictions and userdata checks
+
+## [1.3.2]
+
+### Added
+- **Create** - Added support for IMDS Max Hops
+
+## [1.3.1]
+
+### Added
+- **Create** - Added support for AWS IMDS v2
+
+## [1.3.0]
+
+### Added
+- **Configuration** - Moved to a new dataclass driven configuration methodology
+
+### Removed
+- **YAML Loader** - Removed the yaml_loader file and related methods
+- **Schema** - Removed the dependency on pyschema
+
+## [1.2.0] - 2025-01-30
+
+### Added
+- **S3 sync** - Allow files to be synced to the Forge instance from an S3 bucket
+
+### Changed
+- **Subcommand arguments** - Allowed CLI arguments from any subcommand to be specified for any other
+
 ## [1.1.1] - 2024-12-12
 
 ### Changed
@@ -67,7 +99,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - **Initial commit** - Forge source code, unittests, docs, pyproject.toml, README.md, and LICENSE files.
 
-[unreleased]: https://github.com/carsdotcom/cars-forge/compare/v1.1.1...HEAD
+[unreleased]: https://github.com/carsdotcom/cars-forge/compare/v1.3.3...HEAD
+[1.3.3]: https://github.com/carsdotcom/cars-forge/compare/v1.3.2...v1.3.3
+[1.3.2]: https://github.com/carsdotcom/cars-forge/compare/v1.3.1...v1.3.2
+[1.3.1]: https://github.com/carsdotcom/cars-forge/compare/v1.3.0...v1.3.1
+[1.3.0]: https://github.com/carsdotcom/cars-forge/compare/v1.2.0...v1.3.0
+[1.2.0]: https://github.com/carsdotcom/cars-forge/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/carsdotcom/cars-forge/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/carsdotcom/cars-forge/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/carsdotcom/cars-forge/compare/v1.0.1...v1.0.2